Added a routes dir. Added a helper to the routes directory. Renamed the routes imports on all files that use the import.
