+++ src/containers/Cart/CheckoutFree.js
@@ -10,7 +10,8 @@ import Col from 'react-bootstrap/Col'
 import Alert from 'react-bootstrap/Alert'
 
 import './Cart.css'
-import * as ROUTES from '../../constants/routes'
+// import * as ROUTES from '../../constants/routes'
+import * as ROUTES from '../../routes/constants/routes'
 import { useFirebase, useStateValue } from '../../hooks'
 
 import AutoSuggest from '../../components/AutoSuggest/AutoSuggest'
