+++ src/containers/Cart/OrderReceived.js
@@ -9,7 +9,8 @@ import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 
 import './Cart.css'
-import * as ROUTES from '../../constants/routes'
+// import * as ROUTES from '../../constants/routes'
+import * as ROUTES from '../../routes/constants/routes'
 import { useFirebase, useFirebaseAuthentication } from '../../hooks'
