+++ src/containers/Products/Discover/DiscoverHomeowner.js
@@ -10,7 +10,8 @@ import '../Products.css'
 
 import ProductBox from '../../../components/Product/ProductBox' 
 import Details from '../../../components/Product/Details'
-import * as ROUTES from '../../../constants/routes'
+import * as ROUTES from '../../../routes/constants/routes'
+// import * as ROUTES from '../../../constants/routes'
 
 import { connect } from 'react-redux'
 import { addToCart } from '../../../redux/actions/cartActions'
