+++ src/containers/Products/Reduce/Reduce.js
@@ -10,7 +10,8 @@ import '../Products.css'
 import { Button, TopSection } from '../StyledComponents'
 import sampleReduce from '../../../assets/images/Reduce_Eliminate.svg'
 import SecondRow from '../../Header/SecondRow'
-import * as ROUTES from '../../../constants/routes'
+import * as ROUTES from '../../../routes/constants/routes'
+// import * as ROUTES from '../../../constants/routes'
 
 import ReduceOptions from '../../Options/ReduceOptions'
 import Faq from './FAQ/Faq'
