+++ src/routes/helpers/RouteHelper.js
@@ -1,0 +1,5 @@
+export const convertToProductPathName = (cartItemTitle) => {
+  let productPathName = cartItemTitle.replace(/\s+/g, '-')
+  productPathName = productPathName.replace(/\+/g, '-plus').toLowerCase()
+  return productPathName
+}
